refactor: create test directories
