+++ client/src/components/screens/Patient/components/FamilyTab/tests/family.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import AppTest from 'AppTest';
 
-import FamilyTab from './index';
+import FamilyTab from '../index';
 
 const patientSubState = {
   patient: {
